split out cfg(test) modules
